+++ tests/Queries/AbstractPaginatableQueryTest.php
@@ -18,7 +18,7 @@ class AbstractPaginatableQueryTest extends TestCase
             'country'       => 'CAN',
         ];
 
-        $query = new class($criteria, [], 2, 10) extends AbstractPaginatableQuery {};
+        $query = new class ($criteria, [], 2, 10) extends AbstractPaginatableQuery {};
 
         $this->assertEquals(2, $query->page());
         $this->assertEquals(10, $query->perPage());
@@ -27,5 +27,21 @@ class AbstractPaginatableQueryTest extends TestCase
         $this->assertIsArray($query->includes());
         $this->assertCount(2, $query->criteria());
         $this->assertCount(0, $query->orderBy());
+    }
+
+    public function testCanAddIncludes()
+    {
+        $query = new class ([], [], 2, 10) extends AbstractPaginatableQuery {};
+        $query->include('user.address');
+
+        $this->assertEquals(['user.address'], $query->includes());
+    }
+
+    public function testCanAddMetaData()
+    {
+        $query = new class ([], [], 2, 10) extends AbstractPaginatableQuery {};
+        $query->addMetaData(['include' => 'user.address']);
+
+        $this->assertTrue($query->meta()->has('include'));
     }
 }